Move PeerValidatorRunner back to validation package
